修复： 1、新增实习班种及各班种颜色； 2、动态查询每个人设置的权重； 3、优化标签大小及样式； 4、其他细节优化。

+++ WSBackends/database/models.py
@@ -46,7 +46,7 @@ class WorkschedulePersonnelLink(SQLModel, table=True):
 class Personnel(SQLModel, table=True):
     id: int | None = Field(default=None, primary_key=True)
     name: str = Field(index=True)
-    weight: int = Field(default=1.0, index=True)
+    weight: float = Field(default=1.0, index=True)
     hiredate: date
     worknumber: str
     phonenumber: str
